+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.0
+ January 01, 2020
+- final List<T> items;
+- object.setState(() {});
+- appMenu02.color = Colors.deepOrangeAccent;
+
 ## 0.2.0
  December 31, 2020
 - Provided an example app